新增使用详情和使用汇总的实现

+++ src/main/java/work/daqian/myai/common/PageDTO.java
@@ -10,9 +10,9 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PageDTO<T> {
-    private Integer total;
+    private Long total;
     private List<T> data;
-    public static <T> PageDTO<T> of(Integer total, List<T> data) {
+    public static <T> PageDTO<T> of(Long total, List<T> data) {
         return new PageDTO<>(total, data);
     }
